requirements.txt
aiohttp==3.9.4
alabaster==0.7.16
asgiref==3.7.2
Babel==2.14.0
beautifulsoup4==4.12.3
boto3==1.34.64
botocore==1.34.81
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.10
django-allauth==0.61.1
django-bootstrap-v5==1.0.11
django-heroku==0.3.1
django-storages==1.11.1
docutils==0.16
futures==3.0.5
gunicorn==21.2.0
idna==3.6
imagesize==1.4.1
importlib_metadata==7.1.0
Jinja2==3.1.3
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.7.0
s3==3.0.0
s3transfer==0.10.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==5.3.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
whitenoise==6.6.0
xmltodict==0.13.0
zipp==3.18.1
